% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reilly.R
\name{caReilly}
\alias{caReilly}
\title{Create a stars Raster from a Reilly Regular Grid}
\usage{
caReilly(x, mask, xcoords = "COORDX", ycoords = "COORDY", var = "OUTPUT")
}
\arguments{
\item{x}{sf object; output of the \code{reilly} function.}

\item{mask}{sf object; this object is used to clip 
the raster. (optional)}

\item{xcoords}{character; name of the X coordinates field in x.}

\item{ycoords}{character; name of the Y coordinates field in x.}

\item{var}{character; name of the OUTPUT field in x.}
}
\value{
Catchement areas
}
\description{
This function creates a stars raster from a regularly spaced 
Reilly grid (output of the \code{\link{reilly}} function).
}
\examples{
data(hospital)
# Compute Reilly catchment areas from known points (hospital) on a
# grid defined by its resolution
myreilly <- reilly(knownpts = hospital, varname = "capacity",
                   typefct = "exponential", span = 1250, beta = 3,
                   resolution = 200, mask = paris)
}
