% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a Regularly Spaced Points Grid}
\usage{
create_grid(x, res)
}
\arguments{
\item{x}{an sf object, the spatial extent of this object is used to
create the regular grid.}

\item{res}{resolution of the grid (in map units). If it is not set the grid 
will contain around 5000 points.}
}
\value{
The output of the function is an sf object of regularly spaced
points with the extent of x.
}
\description{
This function creates a regular grid of points
from the extent of a given spatial object and a given resolution.
}
\examples{
# Create a grid of paris extent and 200 meters resolution
library(sf)
data(hospital)
mygrid <- create_grid(x = paris, res = 200)
plot(st_geometry(mygrid), cex = 0.1, pch = ".")
plot(st_geometry(paris), border = "red", lwd = 2, add = TRUE)
}
