% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipotential.R
\name{equipotential}
\alias{equipotential}
\title{Contours of Equipotential}
\usage{
equipotential(
  x,
  nclass = 8,
  breaks,
  mask,
  xcoords = "COORDX",
  ycoords = "COORDY",
  var = "OUTPUT"
)
}
\arguments{
\item{x}{sf POINT data.frame; must contain X, Y and OUTPUT fields.}

\item{nclass}{numeric; a number of class.}

\item{breaks}{numeric; a vector of break values.}

\item{mask}{sf POLYGON data.frame; mask used to 
clip contour shapes.}

\item{xcoords}{character; name of the X coordinates field in x.}

\item{ycoords}{character; name of the Y coordinates field in x.}

\item{var}{character; name of the OUTPUT field in x.}
}
\value{
The output is an sf POLYGON data.frame.
The data frame contains four fields: 
id (id of each polygon), min and max (minimum and maximum breaks of the polygon), 
center (central values of classes).
}
\description{
This function creates spatial polygons of contours equipotential from 
a potential points layer
}
\examples{
# Create a grid of paris extent and 200 meters resolution
data(hospital)
# Compute Stewart potentials from known points (hospital) on a
# grid defined by its resolution (200 m)
pot2 <- potential(x = hospital, var = "capacity",
                  fun = "e", span = 1250, beta = 3,
                  res = 200, mask = paris)

equipot <- equipotential(pot, mask = paris)

plot(equipot)
}
