% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{potential}
\alias{potential}
\title{Compute Potentials}
\usage{
potential(
  x,
  y,
  d,
  var,
  fun = "e",
  span,
  beta,
  res,
  mask,
  bypassctrl = FALSE,
  longlat = TRUE
)
}
\arguments{
\item{x}{sf object; this is the set of known observations to 
estimate the potentials from.}

\item{y}{sf object; this is the set of unknown units for which 
the function computes the estimates. Not used when \code{res} is set 
up. (optional)}

\item{d}{matrix; distance matrix between known observations and unknown 
units for which the function computes the estimates. Row names match the row 
names of \code{x} and column names match the row names of 
\code{y}. \code{d} can contain any distance metric (time 
distance or euclidean distance for example). If \code{matdist} is missing, the distance 
matrix is built with \code{\link{create_matrix}}. (optional)}

\item{var}{character; name of the variable in the \code{x} dataframe 
from which potentials are computed. Quantitative variable with no negative values.}

\item{fun}{spatial interaction function. Options are "p"
(pareto, power law) or "e" (exponential).
For pareto the interaction is defined as: (1 + alpha * mDistance) ^ (-beta).
For "exponential" the interaction is defined as:
exp(- alpha * mDistance ^ beta).
The alpha parameter is computed from parameters given by the user
(\code{beta} and \code{span}).}

\item{span}{numeric; distance where the density of probability of the spatial 
interaction function equals 0.5.}

\item{beta}{numeric; impedance factor for the spatial interaction function.}

\item{res}{numeric; resolution of the output grid (in map units). 
If resolution is not set, the grid will contain around 7250 points. (optional)}

\item{mask}{sf object; the spatial extent of this object is used to 
create the regularly spaced points output. (optional)}

\item{bypassctrl}{logical; bypass the distance matrix size control (see 
\code{\link{create_matrix}} Details).}

\item{longlat}{logical; if FALSE, Euclidean distance, if TRUE Great Circle 
(WGS84 ellipsoid) distance.}
}
\value{
Point object with the computed potentials in a new field 
named \code{OUTPUT}.
}
\description{
This function computes the potentials as defined by J.Q. Stewart (1941).
}
\examples{
# Create a grid of paris extent and 200 meters resolution
data(hospital)
g <- create_grid(x = paris, res = 200)

# Create a distance matrix between known points (hospital) and points 
# of the grid
d <- create_matrix(x = hospital, y = g)

# Compute Stewart potentials from known points (hospital) on a given
# grid (g) using a given distance matrix (d)
pot <- potential(x = hospital, y = g,
                 d = d, var = "capacity",
                 fun = "e", span = 1000,
                 beta = 3, mask = paris)

# Compute Stewart potentials from known points (hospital) on a
# grid defined by its resolution
pot2 <- potential(x = hospital, var = "capacity",
                  fun = "e", span = 1250, beta = 3,
                  res = 200, mask = paris)

# The two methods have the same result
identical(pot, pot2)
# the function output a sf data.frame
class(pot)
# Computed values
summary(pot$OUTPUT)
}
\references{
STEWART J.Q. (1942) "Measure of the influence of a population at a distance", Sociometry, 5(1): 63-71.
}
