% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huff.R
\name{starsHuff}
\alias{starsHuff}
\title{Create a Stars Raster from a Huff sf Object}
\usage{
starsHuff(x, mask, xcoords = "COORDX", ycoords = "COORDY", var = "OUTPUT")
}
\arguments{
\item{x}{sf object; output of the \code{huff} function.}

\item{mask}{sf POLYGON data.frame; mask used to clip 
the stars raster. (optional)}

\item{xcoords}{character; name of the X coordinates field in x.}

\item{ycoords}{character; name of the Y coordinates field in x.}

\item{var}{character; name of the OUTPUT field in x.}
}
\value{
Stars raster of catchment areas values.
}
\description{
This function creates a stars raster from a regularly spaced 
Huff grid (output of the \code{\link{huff}} function).
}
\examples{
library(stars)
library(sf)
data(hospital)
# Compute Huff catchment areas from known points (hospital) on a
# grid defined by its resolution
myhuff <- huff(knownpts = hospital, varname = "capacity",
               typefct = "exponential", span = 750, beta = 2,
               resolution = 100, mask = paris)
# Create a raster of huff values
myhuffstars <- starsHuff(x = myhuff, mask = paris)
plot(myhuffstars, nbreaks = 11, breaks = "equal", 
  col = hcl.colors(10, "BrBG"), 
  main = "Huff Catchment Areas", reset = FALSE)
plot(st_geometry(paris), lwd = 2, add= TRUE)
}
\seealso{
\link{huff}
}
