% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipotential.R
\name{equipotential}
\alias{equipotential}
\title{Contours of Equipotential}
\usage{
equipotential(
  x,
  nclass = 8,
  breaks,
  mask,
  xcoords = "COORDX",
  ycoords = "COORDY",
  var = "OUTPUT"
)
}
\arguments{
\item{x}{an sf object of regularly spaced points. It must contain "X", "Y" 
and "OUTPUT" fields.}

\item{nclass}{a number of class.}

\item{breaks}{a vector of break values.}

\item{mask}{an sf object of polygons or multipolygons. /code{mask} is  used 
to clip polygons of contours equipotential.}

\item{xcoords}{name of the X coordinates field in \code{x}.}

\item{ycoords}{name of the Y coordinates field in \code{x}.}

\item{var}{name of the OUTPUT field in \code{x}.}
}
\value{
The output is an sf object of polygons. The data frame contains four 
fields: id (id of each polygon), min and max (minimum and maximum breaks of 
the polygon) and center (central values of classes).
}
\description{
This function creates spatial polygons of contours equipotential from
a potential points layer
}
\examples{
library(sf)
# Create a grid of paris extent and 200 meters resolution
data(hospital)
# Compute Stewart potentials from known points (hospital) on a
# grid defined by its resolution (200 m)
pot <- potential(
  x = hospital, var = "capacity",
  fun = "e", span = 1250, beta = 3,
  res = 150, mask = paris
)

equipot <- equipotential(pot, mask = paris)

plot(st_geometry(equipot), 
     col = hcl.colors(nrow(equipot), "cividis"))
}
