% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{potential}
\alias{potential}
\title{Compute Potentials}
\usage{
potential(
  x,
  y,
  d,
  var,
  fun = "e",
  span,
  beta,
  res,
  mask,
  bypassctrl = FALSE,
  longlat = TRUE
)
}
\arguments{
\item{x}{an sf object, this is the set of known observations to
estimate the potentials from.}

\item{y}{an sf object, this is the set of unknown units for which
the function computes the estimates. Not used when \code{res} is set
up.}

\item{d}{a distance matrix between known observations and unknown
units for which the function computes the estimates. Row names match the row
names of \code{x} and column names match the row names of
\code{y}. \code{d} can contain any distance metric (time
distance or euclidean distance for example). If \code{d} is missing, 
the distance matrix is built with \code{\link{create_matrix}}.}

\item{var}{name of the variable in \code{x} from which potentials are 
computed. Use only quantitative variable with no negative values.}

\item{fun}{spatial interaction function. Options are "p"
(pareto, power law) or "e" (exponential).
For pareto the interaction is defined as: (1 + alpha * mDistance) ^ (-beta).
For "exponential" the interaction is defined as:
exp(- alpha * mDistance ^ beta).
The alpha parameter is computed from parameters given by the user
(\code{beta} and \code{span}).}

\item{span}{distance where the density of probability of the spatial
interaction function equals 0.5.}

\item{beta}{impedance factor for the spatial interaction function.}

\item{res}{resolution of the output grid (in map units).
If \code{res} is not set, the grid will contain around 5000 points.}

\item{mask}{an sf object, the spatial extent of this object is used to
create the regularly spaced points output if \code{y} is not set.}

\item{bypassctrl}{if TRUE, bypass the distance matrix size control 
(see Details).}

\item{longlat}{if FALSE, the Euclidean distance is used, if TRUE Great Circle
(WGS84 ellipsoid) distance is used.}
}
\value{
sf object with the computed potentials in a new field
named \code{OUTPUT}.
}
\description{
This function computes potentials as defined 
by J.Q. Stewart (1941).
}
\examples{
# Create a grid of paris extent and 200 meters resolution
data(hospital)
g <- create_grid(x = paris, res = 200)

# Create a distance matrix between known points (hospital) and points
# of the grid
d <- create_matrix(x = hospital, y = g)

# Compute Stewart potentials from known points (hospital) on a given
# grid (g) using a given distance matrix (d)
pot <- potential(
  x = hospital, y = g,
  d = d, var = "capacity",
  fun = "e", span = 1250,
  beta = 3, mask = paris
)

# Compute Stewart potentials from known points (hospital) on a
# grid defined by its resolution
pot2 <- potential(
  x = hospital, var = "capacity",
  fun = "e", span = 1250, beta = 3,
  res = 200, mask = paris
)

# The two methods have the same result
identical(pot, pot2)
# the function output a sf data.frame
class(pot)
# Computed values
summary(pot$OUTPUT)
}
\references{
STEWART, JOHN Q. 1941. "An Inverse Distance Variation for Certain Social 
Influences." \emph{Science} 93 (2404): 89–90. 
\url{https://doi.org/10.1126/science.93.2404.89}.
}
